requirements.txt
# Core dependencies for food recommendation system
langchain>=0.1.0
langchain-community>=0.0.10
langchain-chroma>=0.2.0
langchain-huggingface>=0.0.1
langchain-groq>=0.1.0
supabase>=2.7.0
sentence-transformers>=2.6.0
huggingface-hub>=0.23.0
chromadb>=0.4.15
python-dotenv>=1.0.0
numpy>=1.24.3
torch>=2.1.0
transformers>=4.21.0